+++ src/main/java/machines/interfaces/NumberCompareMachine.java
@@ -3,6 +3,7 @@ package machines.interfaces;
 import java.util.List;
 
 public interface NumberCompareMachine extends Machine {
+
 	void inputWinningNums();
 
 	int compare(List<Integer> userNums);
