+++ src/main/java/machines/interfaces/RandomNumsGenerator.java
@@ -2,6 +2,6 @@ package machines.interfaces;
 
 import java.util.List;
 
-public interface RandomNumsGenerator extends Machine{
- 	List<?> createRandomNums();
+public interface RandomNumsGenerator extends Machine {
+	List<?> createRandomNums();
 }
